Stylecop conformance, remove old binder.
Stylecop conformance.
